requirements-dev.txt
pytest
pytest-cov
pytest-django
model_bakery
drf-spectacular

django
djangorestframework
django-rest-passwordreset
psycopg2-binary
pyyaml
ujson
requests
python-dotenv
celery
django-silk